chore: extract localstorage utils to its own file
