+++ src/lib/utils/index.ts
@@ -4,7 +4,6 @@ import { initial } from 'lodash';
 import { Ref } from 'vue';
 import { Path } from 'vue-i18n';
 
-import pkg from '@/../package.json';
 import { NATIVE_ASSET_ADDRESS } from '@/constants/tokens';
 import { POOLS } from '@/constants/pools';
 
@@ -16,49 +15,6 @@ export async function sleep(time) {
   return new Promise(resolve => {
     setTimeout(resolve, time);
   });
-}
-
-function lsAddVersion(value: any, version: string) {
-  return {
-    data: value,
-    _version: version,
-  };
-}
-
-function lsGetKey(key: string) {
-  return `${pkg.name}.${key}`;
-}
-
-export function lsSet(key: string, value: any, version?: string) {
-  const data = version != null ? lsAddVersion(value, version) : value;
-
-  return localStorage.setItem(lsGetKey(key), JSON.stringify(data));
-}
-
-export function lsGet<T = any>(
-  key: string,
-  defaultValue: any = null,
-  version?: string
-): T {
-  const rawValue = localStorage.getItem(lsGetKey(key));
-
-  if (rawValue != null) {
-    try {
-      const value = JSON.parse(rawValue);
-      if (version != null) {
-        return value._version === version ? value.data : defaultValue;
-      }
-      return value;
-    } catch (e) {
-      return defaultValue;
-    }
-  }
-
-  return defaultValue;
-}
-
-export function lsRemove(key: string) {
-  return localStorage.removeItem(lsGetKey(key));
 }
 
 export function getCurrentTs() {
