+++ src/pages/recovery-exit/UserInvestedInAffectedPoolAlert.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import useNetwork from '@/composables/useNetwork';
-import { lsGet, lsSet } from '@/lib/utils';
+import { lsGet, lsSet } from '@/lib/utils/localstorage';
 import { useUserIsDepositedInAffectedPool } from '@/pages/recovery-exit/useUserIsDepositedInAffectedPool';
 import LS_KEYS from '@/constants/local-storage.keys';
